+++ src/main/java/schauweg/smoothswapping/config/Config.java
@@ -1,5 +1,7 @@
 package schauweg.smoothswapping.config;
 
+import schauweg.smoothswapping.Vec2;
+
 import java.util.ArrayList;
 import java.util.List;
