+++ src/main/java/schauweg/smoothswapping/config/ConfigScreen.java
@@ -9,6 +9,7 @@ import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.option.SimpleOption;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.text.Text;
+import schauweg.smoothswapping.Vec2;
 
 import java.util.List;
